fixes and countdown
