+++ src/routes/+layout.svelte
@@ -2,202 +2,85 @@
 	import Background from '$lib/components/background.svelte';
 	import '@fontsource/roboto';
 	import '../app.css';
-	import logo from '$lib/images/logo.svg';
-	import writing from '$lib/images/writing.svg';
-	import menu from '$lib/images/menu.svg';
-	import X from '$lib/images/x.svg';
-	import telegram from '$lib/images/telegram.svg';
-	import discord from '$lib/images/discord.svg';
-	import coingecko from '$lib/images/coingecko.svg';
 	import { page } from '$app/state';
+	import Footer from '$lib/components/footer.svelte';
+	import Header from '$lib/components/header.svelte';
+	import { onMount } from 'svelte';
 
 	let { children } = $props();
 
-	let menuDiv: HTMLDivElement;
+	let isSwap = $derived(page.url.pathname.match(/swap/i) !== null);
 
-	function toggleMenu() {
-		if (menuDiv.classList.contains('open')) menuDiv.classList.remove('open');
-		else menuDiv.classList.add('open');
+	let target = new Date("2024-12-27T12:00:00Z")
+	let time = $state(target.valueOf() - new Date().valueOf())
+
+
+	const MILLIS_PER_HOUR = 60*60*1000
+	const MILLIS_PER_MINUTE = 60*1000
+	function splitTimeDiff(time: number) {
+		let hours = Math.floor(time / MILLIS_PER_HOUR)
+		let t = time - hours * MILLIS_PER_HOUR
+		let minutes = Math.floor(t / MILLIS_PER_MINUTE)
+		t -= minutes * MILLIS_PER_MINUTE
+		let seconds = Math.floor(t / 1000)
+		return {
+			hours: twoDigits(hours), 
+			minutes: twoDigits(minutes), 
+			seconds: twoDigits(seconds)
+		}
+	}
+	function twoDigits(num: number) {
+		let n = num.toString();
+		return n.length === 2? n : '0'+n
 	}
 
-	let isSwap = $derived(page.url.pathname.match(/swap/i) !== null);
+	let timeDiff = $derived(splitTimeDiff(time))
+
+	onMount(() => {
+		const interval = setInterval(() => {
+			time = target.valueOf() - new Date().valueOf();
+		}, 1000);
+
+		return () => {
+			clearInterval(interval);
+		};
+	});
 </script>
 
 <div class="overflow-hidden relative bg-black min-h-screen flex flex-col justify-between">
-	<div id="positioner" class="absolute w-full z-10">
-		<div id="header" class="allcenter bg-transparent text-white py-3 px-7">
-			<div id="header-content" class="allcenter max-w-[1200px] w-full">
-				<a href="/" id="brand" class="allcenter">
-					<img src={logo} alt="logo" class="h-[50px]" />
-					<img src={writing} alt="choccyswap" class="h-[30px] px-3" />
-				</a>
-				<div class="max-[599px]:hidden allcenter ml-auto">
-					<div id="links" class="allcenter ml-auto space-x-8 font-medium">
-						<a href="/" class=""> Stats </a>
-						<a href="/" class=""> Pools </a>
-					</div>
-					{#if isSwap}
-						<button id="swapbutton" class="py-1.5 px-7 font-semibold ml-8"> Connect Wallet </button>
-					{:else}
-						<a href="/swap" id="swapbutton" class="py-1.5 px-7 font-semibold ml-8"> Swap Now </a>
-					{/if}
-				</div>
-
-				<div class="min-[600px]:hidden ml-auto relative">
-					<button onclick={toggleMenu}>
-						<img src={menu} alt="menu" class="text-white" />
-					</button>
-					
-					<div
-						bind:this={menuDiv}
-						id="topbarmenu"
-						class="min-[600px]:!hidden absolute top-full right-0 p-3 gap-3 w-[500%]"
-					>
-						<a href="/" class=""> Stats </a>
-						<a href="/" class=""> Pools </a>
-	
-						{#if isSwap}
-							<button id="swapbutton" class="py-1.5 px-7 font-semibold"> Connect Wallet </button>
-						{:else}
-							<a href="/swap" id="swapbutton" class="py-1.5 px-7 font-semibold"> Swap Now </a>
-						{/if}
-					</div>
-				</div>
-			</div>
-		</div>
-	</div>
+	<!--Header {isSwap} />
 
 	<Background>
 		{@render children()}
 	</Background>
 
 	{#if !isSwap}
-		<div class="relative">
-			<div id="footer" class="flex flex-col items-center w-full text-white bg-transparent">
-				<span class="c ctopright"></span>
-				<span class="c ctopleft"></span>
-				<div id="socials" class="allcenter space-x-8 h-[6cm]">
-					<a href="https://x.com/ChoccySwap" target="_blank">
-						<img src={X} alt="X twitter" />
-					</a>
-					<a href="https://t.me/ChoccySwapCommunity" target="_blank">
-						<img src={telegram} alt="telegram" />
-					</a>
-					<a href="https://discord.com/invite/cunRKsgbBz" target="_blank">
-						<img src={discord} alt="discord" />
-					</a>
-					<a href="https://www.coingecko.com/en/coins/choccyswap" target="_blank">
-						<img src={coingecko} alt="coingecko" />
-					</a>
-				</div>
-				<img src={writing} alt="ChoccySwap" />
-			</div>
-		</div>
-	{/if}
+		<Footer/>
+	{/if}-->
+	<p class="allcenter">{timeDiff.hours}:{timeDiff.minutes}:{timeDiff.seconds}</p>
+	<p class="allcenter">Phase 0/4</p>
 </div>
 
-<style lang="less">
-	#swapbutton {
-		background: linear-gradient(276.31deg, #ed32bf -10.17%, #b5178e 97.19%);
-		border-radius: 32px;
-		backdrop-filter: blur(5.538px);
-		border: 2.34px solid #b5178e;
-
-		&:hover {
-			border: 2.34px solid rgba(255, 255, 255, 0.326);
+<style>
+	div {
+		display: flex;
+		flex-direction: column;
+		align-items: center;
+		justify-content: center;
+	}
+	p {
+		color: #b5178e;
+		font-weight: 900;
+		font-size: 90pt;
+	}
+	@media(max-width: 520px) {
+		p {
+			font-size: 50pt !important;
 		}
 	}
-
-	#header::before {
-		content: '';
-		position: absolute;
-		left: 0;
-		top: 0;
-		width: 100%;
-		height: 100%;
-		z-index: -1;
-		backdrop-filter: blur(5.538px);
-	}
-
-	#topbarmenu {
-		display: none;
-
-		&::before {
-			content: '';
-			position: absolute;
-			left: 0;
-			top: 0;
-			width: 100%;
-			height: 100%;
-			z-index: -1;
-			background-color: #ed32bf22;
-			border-radius: 5mm;
-			backdrop-filter: blur(5.538px);
-		}
-	}
-
-	#footer {
-		border-radius: 5mm 5mm 0 0;
-		overflow: hidden;
-		position: relative;
-		z-index: 1;
-		height: calc(7cm + min(100vw, 30cm) / 5);
-		&::before {
-			content: '';
-			position: absolute;
-			left: 0;
-			top: 0;
-			width: 100%;
-			height: 100%;
-			z-index: -1;
-			backdrop-filter: blur(20px);
-		}
-		& > img {
-			max-width: 30cm;
-			width: 100%;
-			padding: 0 1rem;
-			position: absolute;
-			bottom: 1cm;
-		}
-	}
-
-	:global(body) {
-		font-family: Roboto, sans-serif;
-	}
-
-	.c {
-		position: absolute;
-		height: 200px;
-		width: 200px;
-		border-radius: 100%;
-		filter: blur(120px);
-
-		&topright {
-			height: 400px;
-			width: 400px;
-			transform: rotate(-45deg);
-
-			background-image: linear-gradient(
-				90deg,
-				rgba(147, 238, 255, 0.6) 13%,
-				rgba(255, 2, 209, 0.6) 85%
-			);
-			top: -100px;
-			right: -100px;
-		}
-
-		&topleft {
-			height: 400px;
-			width: 400px;
-			transform: rotate(45deg);
-
-			background-image: linear-gradient(
-				-90deg,
-				rgba(147, 238, 255, 0.6) 13%,
-				rgba(255, 2, 209, 0.6) 85%
-			);
-			top: -100px;
-			left: -100px;
+	@media(max-width: 780px) {
+		p {
+			font-size: 70pt;
 		}
 	}
 </style>
